+++ src/main/java/com/brunch/server/book/message/SuccessMessage.java
@@ -9,6 +9,7 @@ public enum SuccessMessage {
 
     SUCCESS_GET_VIEWED_BOOKS("최근 읽었던 책 목록 조회 성공"),
     SUCCESS_GET_DETAILED_BOOKS("책 상세 조회 성공"),
+    SUCCESS_GET_BANNER("배너 리스트 조회 성공")
     ;
 
     private final String message;
